discographyのディレクトリ構成を変更
