+++ app/pages/discography/[id].vue
@@ -1,0 +1,7 @@
+<template>
+
+</template>
+
+<script setup>
+
+</script>
